+++ atlas/modules/helpers/generate_routes.py
@@ -31,8 +31,7 @@ class GenerateRoutes:
                 if method in constants.VALID_METHODS:
                     # Ignore Exclude URLs
                     op_key = f"{method.upper()} {path}"
-                    identifier = config.get(constants.OPERATION, utils.operation_id_name(path, method))
-                    identifier = re.sub(r'[.-]', '_', identifier)
+                    identifier = re.sub(r'[.-]', '_', utils.operation_id_name(path, method))
                     routes.append(f"{identifier.upper()} = '{op_key}'")
 
         return "\n".join(routes)